店铺列表接口，以及数据填充

+++ ZLead/Tool/Net/NetManager.h
@@ -25,7 +25,7 @@ typedef NS_ENUM(NSInteger,HttpRequestType) {
 typedef void(^failWithErrorBlock)(NSError *error);
 typedef void(^successfulBlock)(void);
 typedef void(^resultBlock)(id result);
-typedef void(^successBlock)(NSArray *dataList, int total);
+typedef void(^successBlock)(NSArray *dataList, NSInteger total);
 
 @interface NetManager : NSObject
